pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lanrtr"
version = "0.1"
authors = [
  { name="Drew Gochenaur", email="junkyalleycat@raincity.io" },
]
requires-python = ">=3.8"
dependencies = [
  "scapy>=2.5.0",
  "netifaces",
  "lru-dict"
]

[project.scripts]
lanrtr-rtsol = "lanrtr.rt_sol:entry"
lanrtr-nbradv = "lanrtr.nbr_adv:entry"
